refactor(query): extract system-table builders into a submodule + WILDCARD guard
executor/mod.rs was 4458 lines. Move the ten build_*_table builders (+ the
entries row helper) into executor/system_tables.rs (mod.rs -> 3712 lines); the
get_builtin_table dispatcher still calls them. Add a WILDCARD_COLUMNS alignment
test: SELECT * header == WILDCARD_COLUMNS and every row width == its len, so the
name list and the evaluation.rs value-push cannot silently drift. Pure code move
+ guard. Architecture-roadmap item [M/BR4].
